Some minor cleanup
