+++ src/VisualStudio/Next/Remote/RemoteHostClientServiceFactory.cs
@@ -1,0 +1,27 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System.Composition;
+using Microsoft.CodeAnalysis.Diagnostics;
+using Microsoft.CodeAnalysis.Host;
+using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.VisualStudio.LanguageServices.Implementation.Remote;
+
+namespace Microsoft.VisualStudio.LanguageServices.Remote
+{
+    [ExportWorkspaceServiceFactory(typeof(IRemoteHostClientService)), Shared]
+    internal partial class RemoteHostClientServiceFactory : IWorkspaceServiceFactory
+    {
+        private readonly IDiagnosticAnalyzerService _analyzerService;
+
+        [ImportingConstructor]
+        public RemoteHostClientServiceFactory(IDiagnosticAnalyzerService analyzerService)
+        {
+            _analyzerService = analyzerService;
+        }
+
+        public IWorkspaceService CreateService(HostWorkspaceServices workspaceServices)
+        {
+            return new RemoteHostClientService(workspaceServices.Workspace, _analyzerService);
+        }
+    }
+}
