+++ src/VisualStudio/Next/ServicesVisualStudio.Next.csproj
@@ -82,7 +82,8 @@
       <Link>Shared\WellKnownServiceHubServices.cs</Link>
     </Compile>
     <Compile Include="Diagnostics\OutOfProcCompilerDiagnosticAnalyzer.cs" />
-    <Compile Include="Remote\RemoteHostClientService.cs" />
+    <Compile Include="Remote\RemoteHostClientServiceFactory.RemoteHostClientService.cs" />
+    <Compile Include="Remote\RemoteHostClientServiceFactory.cs" />
     <Compile Include="Remote\JsonRpcClient.cs" />
     <Compile Include="Remote\RemoteHostOptions.cs" />
     <Compile Include="Remote\ServiceHubRemoteHostClient.JsonRpcSnapshot.cs" />
